register update and test
Fixed some confused ocr1a/b register settings and tested it on actual hardware.
It functions correctly as a one-shot.  The minimum shot time is 66nS, as measured on a scope.  However, handing it 60,000 does not result in a 90% duty cycle, as I would have expected.  Still working on it.